requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
PyJWT==2.10.1
python-dotenv==1.0.0
requests==2.32.4
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3

